code/requirements.txt
joblib==1.4.2
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
PyYAML==6.0.2
scikit_learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
tqdm==4.67.1
chardet==5.2.0
